src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/primitive/ArrayPrimitiveBooleanTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayPrimitiveBooleanTextParser extends ArrayTextParser {

    public ArrayPrimitiveBooleanTextParser() {
        super(false);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getText());
    }

}
